spritesheet_animation.cc:L134/revert changes, upcast 'rows'
to fix "Multiplication result converted to larger type" (codeql)

https://github.com/andy5995/widelands/security/code-scanning/2

+++ src/graphic/animation/spritesheet_animation.cc
@@ -131,8 +131,8 @@ SpriteSheetAnimation::SpriteSheetMipMapEntry::frame_textures(bool return_playerc
 	std::vector<std::unique_ptr<const Texture>> result;
 	const Rectf rect(Vector2f::zero(), width(), height());
 	if (!return_playercolor_masks || has_playercolor_masks) {
-		const int no_of_frames = rows * columns;
-		for (int i = 0; i < no_of_frames; ++i) {
+		const size_t no_of_frames = static_cast<size_t>(rows) * columns;
+		for (size_t i = 0; i < no_of_frames; ++i) {
 			std::unique_ptr<Texture> texture(new Texture(width(), height()));
 
 			const int column = i % columns;
